Remove ApiRequestException

+++ back_end/src/main/java/com/springboot/shopping/service/impl/UserServiceImpl.java
@@ -5,7 +5,6 @@ import java.util.Collection;
 import java.util.List;
 import java.util.Optional;
 
-import org.springframework.http.HttpStatus;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.User;
 import org.springframework.security.core.userdetails.UserDetails;
@@ -18,11 +17,11 @@ import org.springframework.validation.BindingResult;
 import com.springboot.shopping.dto.PasswordResetRequest;
 import com.springboot.shopping.dto.user.UserRequest;
 import com.springboot.shopping.dto.user.UserResponse;
-import com.springboot.shopping.exception.ApiRequestException;
 import com.springboot.shopping.exception.InputFieldException;
-import com.springboot.shopping.exception.PasswordConfirmationException;
-import com.springboot.shopping.exception.UserNotFoundException;
-import com.springboot.shopping.exception.UserRoleExistException;
+import com.springboot.shopping.exception.auth.PasswordConfirmationException;
+import com.springboot.shopping.exception.role.RoleNotFoundException;
+import com.springboot.shopping.exception.user.UserNotFoundException;
+import com.springboot.shopping.exception.user.UserRoleExistException;
 import com.springboot.shopping.mapper.CommonMapper;
 import com.springboot.shopping.model.Role;
 import com.springboot.shopping.model.UserEntity;
@@ -43,34 +42,40 @@ public class UserServiceImpl implements UserService, UserDetailsService {
 
 	@Override
 	public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-		Optional<UserEntity> userEntity = userRepository.findByUsername(username);
-		if (!userEntity.isPresent()) {
+		Optional<UserEntity> userFromDb = userRepository.findByUsername(username);
+		if (userFromDb.isEmpty()) {
 			throw new UserNotFoundException();
 		}
 		Collection<SimpleGrantedAuthority> Authorities = new ArrayList<>();
-		userEntity.get().getRoles().forEach(role -> Authorities.add(new SimpleGrantedAuthority(role.getName())));
-		return new User(userEntity.get().getUsername(), userEntity.get().getPassword(), Authorities);
+		userFromDb.get().getRoles().forEach(role -> Authorities.add(new SimpleGrantedAuthority(role.getName())));
+		return new User(userFromDb.get().getUsername(), userFromDb.get().getPassword(), Authorities);
 	}
 
 	@Override
 	public UserResponse findUserById(Long userId) {
-		UserEntity userFromDB = userRepository.findById(userId)
-				.orElseThrow(() -> new ApiRequestException("User not found!", HttpStatus.NOT_FOUND));
-		return commonMapper.convertToResponse(userFromDB, UserResponse.class);
+		Optional<UserEntity> userFromDb = userRepository.findById(userId);
+		if (userFromDb.isEmpty()) {
+			throw new UserNotFoundException();
+		}
+		return commonMapper.convertToResponse(userFromDb.get(), UserResponse.class);
 	}
 
 	@Override
 	public UserResponse findUserByUsername(String username) {
-		UserEntity userFromDB = userRepository.findByUsername(username)
-				.orElseThrow(() -> new ApiRequestException("User not found!", HttpStatus.NOT_FOUND));
-		return commonMapper.convertToResponse(userFromDB, UserResponse.class);
+		Optional<UserEntity> userFromDb = userRepository.findByUsername(username);
+		if (userFromDb.isEmpty()) {
+			throw new UserNotFoundException();
+		}
+		return commonMapper.convertToResponse(userFromDb, UserResponse.class);
 	}
 
 	@Override
 	public UserEntity findUserByUsernameReturnUserEntity(String username) {
-		UserEntity userFromDB = userRepository.findByUsername(username)
-				.orElseThrow(() -> new ApiRequestException("User not found!", HttpStatus.NOT_FOUND));
-		return userFromDB;
+		Optional<UserEntity> userFromDb = userRepository.findByUsername(username);
+		if (userFromDb.isEmpty()) {
+			throw new UserNotFoundException();
+		}
+		return userFromDb.get();
 	}
 
 	@Override
@@ -84,20 +89,24 @@ public class UserServiceImpl implements UserService, UserDetailsService {
 			throw new InputFieldException(bindingResult);
 		}
 		UserEntity newUserInfo = commonMapper.convertToEntity(userRequest, UserEntity.class);
-		UserEntity userFromDb = userRepository.findByUsername(username)
-				.orElseThrow(() -> new ApiRequestException("Username not found.", HttpStatus.NOT_FOUND));
-		userFromDb.setFirstName(newUserInfo.getFirstName());
-		userFromDb.setLastName(newUserInfo.getLastName());
-		userFromDb.setAddress(newUserInfo.getAddress());
-		userFromDb.setPhoneNumber(newUserInfo.getPhoneNumber());
+		Optional<UserEntity> userFromDb = userRepository.findByUsername(username);
+		if (userFromDb.isEmpty()) {
+			throw new UserNotFoundException();
+		}
+		userFromDb.get().setFirstName(newUserInfo.getFirstName());
+		userFromDb.get().setLastName(newUserInfo.getLastName());
+		userFromDb.get().setAddress(newUserInfo.getAddress());
+		userFromDb.get().setPhoneNumber(newUserInfo.getPhoneNumber());
 
-		return commonMapper.convertToResponse(userRepository.save(userFromDb), UserResponse.class);
+		return commonMapper.convertToResponse(userRepository.save(userFromDb.get()), UserResponse.class);
 	}
 
 	@Override
 	public List<UserResponse> deleteUser(Long userId) {
-		userRepository.findById(userId)
-				.orElseThrow(() -> new ApiRequestException("User not found!", HttpStatus.NOT_FOUND));
+		Optional<UserEntity> userFromDb = userRepository.findById(userId);
+		if (userFromDb.isEmpty()) {
+			throw new UserNotFoundException();
+		}
 		userRepository.deleteById(userId);
 
 		return commonMapper.convertToResponseList(userRepository.findAll(), UserResponse.class);
@@ -105,15 +114,19 @@ public class UserServiceImpl implements UserService, UserDetailsService {
 
 	@Override
 	public String addRoleToUser(String username, String roleName) {
-		UserEntity user = userRepository.findByUsername(username)
-				.orElseThrow(() -> new ApiRequestException("User not found!", HttpStatus.NOT_FOUND));
-		Role role = roleRepository.findByname(roleName)
-				.orElseThrow(() -> new ApiRequestException("Role not found!", HttpStatus.NOT_FOUND));
-		if (user.getRoles().contains(role)) {
-			throw new UserRoleExistException("User already has this role !");
+		Optional<UserEntity> userFromDb = userRepository.findByUsername(username);
+		if (userFromDb.isEmpty()) {
+			throw new UserNotFoundException();
 		}
-		user.getRoles().add(role);
-		userRepository.save(user);
+		Optional<Role> roleFromDb = roleRepository.findByname(roleName);
+		if (roleFromDb.isEmpty()) {
+			throw new RoleNotFoundException();
+		}
+		if (userFromDb.get().getRoles().contains(roleFromDb.get())) {
+			throw new UserRoleExistException();
+		}
+		userFromDb.get().getRoles().add(roleFromDb.get());
+		userRepository.save(userFromDb.get());
 		return "Role successfully added.";
 	}
 
@@ -129,10 +142,12 @@ public class UserServiceImpl implements UserService, UserDetailsService {
 		if (password != null && !password.equals(passwordRepeat)) {
 			throw new PasswordConfirmationException("Passwords do not match.");
 		}
-		UserEntity userFromDb = userRepository.findByUsername(username)
-				.orElseThrow(() -> new ApiRequestException("User not found.", HttpStatus.NOT_FOUND));
-		userFromDb.setPassword(passwordEncoder.encode(password));
-		userRepository.save(userFromDb);
+		Optional<UserEntity> userFromDb = userRepository.findByUsername(username);
+		if (userFromDb.isEmpty()) {
+			throw new UserNotFoundException();
+		}
+		userFromDb.get().setPassword(passwordEncoder.encode(password));
+		userRepository.save(userFromDb.get());
 		return "Password successfully changed!";
 
 	}
